move interface to types
