+++ packages/repocop/src/types.ts
@@ -1,4 +1,10 @@
-interface RepoAndStack {
+export interface RepoAndStack {
 	full_name: string;
 	stacks: string[];
+}
+
+export interface RepoAndArchiveStatus {
+	full_name: string;
+	name: string;
+	archived: boolean;
 }
